requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
lightgbm
joblib
mlflow
pytest
httpx
xgboost
gradio
elasticsearch
evidentlyai
datetime
json
pathlib